Remove identity/verification file
