+++ packages/mobile/src/account/Settings.tsx
@@ -39,7 +39,6 @@ import Dialog from 'src/components/Dialog'
 import SessionId from 'src/components/SessionId'
 import { TOS_LINK } from 'src/config'
 import { Namespaces, withTranslation } from 'src/i18n'
-import { revokeVerification } from 'src/identity/actions'
 import { LocalCurrencyCode } from 'src/localCurrency/consts'
 import { getLocalCurrencyCode } from 'src/localCurrency/selectors'
 import DrawerTopBar from 'src/navigator/DrawerTopBar'
@@ -51,9 +50,10 @@ import { restartApp } from 'src/utils/AppRestart'
 import { navigateToURI } from 'src/utils/linking'
 import Logger from 'src/utils/Logger'
 import { toggleFornoMode } from 'src/web3/actions'
+import { revoke } from 'src/verify/reducer'
 
 interface DispatchProps {
-  revokeVerification: typeof revokeVerification
+  revokeVerification: typeof revoke
   setNumberVerified: typeof setNumberVerified
   resetAppOpenedState: typeof resetAppOpenedState
   setAnalyticsEnabled: typeof setAnalyticsEnabled
@@ -104,7 +104,7 @@ const mapStateToProps = (state: RootState): StateProps => {
 }
 
 const mapDispatchToProps = {
-  revokeVerification,
+  revokeVerification: revoke,
   setNumberVerified,
   resetAppOpenedState,
   setAnalyticsEnabled,
