+++ packages/mobile/src/app/saga.ts
@@ -31,7 +31,6 @@ import { getLastTimeBackgrounded, getRequirePinOnAppOpen } from 'src/app/selecto
 import { handleDappkitDeepLink } from 'src/dappkit/dappkit'
 import { CicoProviderNames } from 'src/fiatExchanges/reducer'
 import { appRemoteFeatureFlagChannel, appVersionDeprecationChannel } from 'src/firebase/firebase'
-import { CodeInputType } from 'src/identity/verification'
 import { navigate } from 'src/navigator/NavigationService'
 import { Screens } from 'src/navigator/Screens'
 import { StackParamList } from 'src/navigator/types'
@@ -40,6 +39,7 @@ import { navigateToURI } from 'src/utils/linking'
 import Logger from 'src/utils/Logger'
 import { clockInSync } from 'src/utils/time'
 import { receiveAttestationCode } from 'src/verify/reducer'
+import { CodeInputType } from 'src/verify/saga'
 import { parse } from 'url'
 
 const TAG = 'app/saga'
