+++ packages/mobile/src/identity/saga.ts
@@ -27,9 +27,7 @@ import {
   AddressValidationType,
   e164NumberToAddressSelector,
 } from 'src/identity/reducer'
-import { revokeVerificationSaga } from 'src/identity/revoke'
 import { validateAndReturnMatch } from 'src/identity/secureSend'
-import { reportRevealStatusSaga, startVerificationSaga } from 'src/identity/verification'
 import { Actions as TransactionActions } from 'src/transactions/actions'
 import Logger from 'src/utils/Logger'
 import { fetchDataEncryptionKeyWrapper } from 'src/web3/dataEncryptionKey'
@@ -115,11 +113,6 @@ function* fetchKnownAddresses() {
   }
 }
 
-function* watchVerification() {
-  yield takeLatest(Actions.START_VERIFICATION, startVerificationSaga)
-  yield takeLeading(Actions.REVOKE_VERIFICATION, revokeVerificationSaga)
-}
-
 function* watchContactMapping() {
   yield takeLeading(Actions.IMPORT_CONTACTS, doImportContactsWrapper)
   yield takeEvery(Actions.FETCH_ADDRESSES_AND_VALIDATION_STATUS, fetchAddressesAndValidateSaga)
@@ -137,19 +130,13 @@ function* watchFetchDataEncryptionKey() {
   yield takeLeading(Actions.FETCH_DATA_ENCRYPTION_KEY, fetchDataEncryptionKeyWrapper)
 }
 
-function* watchReportRevealStatus() {
-  yield takeEvery(Actions.REPORT_REVEAL_STATUS, reportRevealStatusSaga)
-}
-
 export function* identitySaga() {
   Logger.debug(TAG, 'Initializing identity sagas')
   try {
-    yield spawn(watchVerification)
     yield spawn(watchContactMapping)
     yield spawn(watchValidateRecipientAddress)
     yield spawn(watchNewFeedTransactions)
     yield spawn(watchFetchDataEncryptionKey)
-    yield spawn(watchReportRevealStatus)
     yield spawn(fetchKnownAddresses)
   } catch (error) {
     Logger.error(TAG, 'Error initializing identity sagas', error)
