+++ packages/mobile/src/identity/smsRetrieval.ts
@@ -1,9 +1,9 @@
 import SmsRetriever from '@celo/react-native-sms-retriever'
 import { eventChannel } from 'redux-saga'
 import { call, put, take } from 'redux-saga/effects'
-import { receiveAttestationMessage } from 'src/identity/actions'
-import { CodeInputType, NUM_ATTESTATIONS_REQUIRED } from 'src/identity/verification'
 import Logger from 'src/utils/Logger'
+import { receiveAttestationCode } from 'src/verify/reducer'
+import { CodeInputType, NUM_ATTESTATIONS_REQUIRED } from 'src/verify/saga'
 
 const TAG = 'identity/smsRetrieval'
 
@@ -22,7 +22,7 @@ export function* startAutoSmsRetrieval() {
   try {
     while (true) {
       const { message } = yield take(autoSmsChannel)
-      yield put(receiveAttestationMessage(message, CodeInputType.AUTOMATIC))
+      yield put(receiveAttestationCode({ message, inputType: CodeInputType.AUTOMATIC }))
     }
   } catch (error) {
     Logger.error(TAG + '@SmsRetriever', 'Error while retrieving code', error)
