+++ packages/mobile/src/verify/VerificationCodeInput.tsx
@@ -6,9 +6,9 @@ import {
 import * as React from 'react'
 import { StyleProp, ViewStyle } from 'react-native'
 import CodeInput, { CodeInputStatus } from 'src/components/CodeInput'
-import { ATTESTATION_CODE_PLACEHOLDER } from 'src/identity/reducer'
-import { AttestationCode } from 'src/identity/verification'
 import Logger from 'src/utils/Logger'
+import { ATTESTATION_CODE_PLACEHOLDER } from 'src/verify/reducer'
+import { AttestationCode } from 'src/verify/saga'
 
 interface Props {
   label: string
