+++ packages/mobile/test/schemas.ts
@@ -2,9 +2,9 @@ import _ from 'lodash'
 import { PincodeType } from 'src/account/reducer'
 import { AppState } from 'src/app/actions'
 import { DEFAULT_DAILY_PAYMENT_LIMIT_CUSD } from 'src/config'
-import { NUM_ATTESTATIONS_REQUIRED } from 'src/identity/verification'
 import { RootState } from 'src/redux/reducers'
 import { idle, KomenciAvailable } from 'src/verify/reducer'
+import { NUM_ATTESTATIONS_REQUIRED } from 'src/verify/saga'
 
 // Default (version -1 schema)
 export const vNeg1Schema = {
